feat: AnimGraph2 update (#93)
* ag2

* 123

* remove fields got nuked by valve

* protobuf

* idk

* 1234

* 1234

* temp fix

* fix windows SetAbsOrigin

* lol

* fix

* patch EnableHammerId gatekeeper in CBaseEntity::Spawn for linux

* fix

* remove fields

* add AimPunchService

* xd

* fix EntityEnhancement

* subtick move

* force finding exact string

* logs

* clean

* deps: nuget packages

---------

+++ Sharp.Core/Managers/Forwards/PlayerKilledForward.cs
@@ -404,7 +404,7 @@ internal sealed unsafe class PlayerKilledForwardParams : PlayerPawnFunctionParam
 
     public int DamageTaken
     {
-        get => Result->DamageDealt;
+        get => (int)Result->DamageDealt;
         set
         {
             CheckDisposed();
@@ -426,7 +426,7 @@ internal sealed unsafe class PlayerKilledForwardParams : PlayerPawnFunctionParam
 
     public int TotalledDamageTaken
     {
-        get => Result->TotalledDamageDealt;
+        get => (int)Result->TotalledDamageDealt;
         set
         {
             CheckDisposed();
